code cleanup

+++ src/main/com/mucommander/commons/file/impl/sevenzip/provider/SevenZip/Compression/LZ/BinTree.java
@@ -140,12 +140,12 @@ public class BinTree extends InWindow
 		{
 			int temp = CrcTable[_bufferBase[cur] & 0xFF] ^ (_bufferBase[cur + 1] & 0xFF);
 			hash2Value = temp & (kHash2Size - 1);
-			temp ^= ((int)(_bufferBase[cur + 2] & 0xFF) << 8);
+			temp ^= (_bufferBase[cur + 2] & 0xFF) << 8;
 			hash3Value = temp & (kHash3Size - 1);
 			hashValue = (temp ^ (CrcTable[_bufferBase[cur + 3] & 0xFF] << 5)) & _hashMask;
 		}
 		else
-			hashValue = ((_bufferBase[cur] & 0xFF) ^ ((int)(_bufferBase[cur + 1] & 0xFF) << 8));
+			hashValue = ((_bufferBase[cur] & 0xFF) ^ ((_bufferBase[cur + 1] & 0xFF) << 8));
 
 		int curMatch = _hash[kFixHashSize + hashValue];
 		if (HASH_ARRAY)
@@ -276,13 +276,13 @@ public class BinTree extends InWindow
 				int temp = CrcTable[_bufferBase[cur] & 0xFF] ^ (_bufferBase[cur + 1] & 0xFF);
 				int hash2Value = temp & (kHash2Size - 1);
 				_hash[hash2Value] = _pos;
-				temp ^= ((int)(_bufferBase[cur + 2] & 0xFF) << 8);
+				temp ^= ((_bufferBase[cur + 2] & 0xFF) << 8);
 				int hash3Value = temp & (kHash3Size - 1);
 				_hash[kHash3Offset + hash3Value] = _pos;
 				hashValue = (temp ^ (CrcTable[_bufferBase[cur + 3] & 0xFF] << 5)) & _hashMask;
 			}
 			else
-				hashValue = ((_bufferBase[cur] & 0xFF) ^ ((int)(_bufferBase[cur + 1] & 0xFF) << 8));
+				hashValue = ((_bufferBase[cur] & 0xFF) ^ ((_bufferBase[cur + 1] & 0xFF) << 8));
 
 			int curMatch = _hash[kFixHashSize + hashValue];
 			_hash[kFixHashSize + hashValue] = _pos;
